Removed Progressive Web App Flag
"display": "standalone" tells browsers that the site is a PWA so changed to "display": "browser"

PWAs allow websites to be "installed" on mobile devices and provide an app-like experience. Before this change, it would prompt visitors to "install" the website as a standalone application. It was working too, if I ever wanted to enable this again in the future.